\name{Pythogorus-package}
\alias{Pythogorus-package}
\alias{Pythogorus}
\docType{package}
\title{
\packageTitle{Pythogorus}
}
\description{
\packageDescription{Pythogorus}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Pythogorus}
\packageIndices{Pythogorus}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Pythogorus}

Maintainer: \packageMaintainer{Pythogorus}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
